requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
psycopg[binary]
python-jose[cryptography]
passlib[bcrypt]
python-multipart
email-validator
pytest
httpx
pytest-asyncio
requests
psycopg2-binary